Create README.md
I am a certified Junior QA Tester with experience in digital environments and skills in software analysis, basic automation and handling of technical tools. My approach is based on quality, creativity and problem solving. With a solid foundation in design and communication, I offer a unique perspective to ensure optimal user experiences. I seek to contribute to dynamic teams that promote innovation and excellence.